// src/api/openapi/auto-generated/models/MeshResultSetOptions.ts
export interface MeshResultSetOptions {
    /**
     * 
     * @type {string}
     * @memberof MeshResultSetOptions
     */
    dataVersion?: MeshResultSetOptionsDataVersionEnum;
    /**
     * 
     * @type {Array<string>}
     * @memberof MeshResultSetOptions
     */
    scenarios?: Array<MeshResultSetOptionsScenariosEnum>;
    /**
     * 
     * @type {Array<number>}
     * @memberof MeshResultSetOptions
     */
    years?: Array<MeshResultSetOptionsYearsEnum>;
    /**
     * 
     * @type {Array<FloodType>}
     * @memberof MeshResultSetOptions
     */
    perils?: Array<FloodType>;
    /**
     * 
     * @type {FloodDefenseOptions}
     * @memberof MeshResultSetOptions
     */
    floodDefense?: FloodDefenseOptions;
    /**
     * 
     * @type {GridOptions}
     * @memberof MeshResultSetOptions
     */
    mesh?: GridOptions;
    /**
     * 
     * @type {string}
     * @memberof MeshResultSetOptions
     */
    type: MeshResultSetOptionsTypeEnum;
}


/**
 * @export
 */
export const MeshResultSetOptionsDataVersionEnum = {
    _26: '2.6',
    _262: '2.6.2',
    _3: '3',
    _30: '3.0',
    _300: '3.0.0',
    _31: '3.1',
    _310: '3.1.0',
    _32: '3.2',
    _320: '3.2.0',
    _330: '3.3.0',
    unknownDefaultOpenApi: '11184809'
} as const;
export type MeshResultSetOptionsDataVersionEnum = typeof MeshResultSetOptionsDataVersionEnum[keyof typeof MeshResultSetOptionsDataVersionEnum];

/**
 * @export
 */
export const MeshResultSetOptionsScenariosEnum = {
    baseline: 'baseline',
    ssp126: 'ssp126',
    ssp245: 'ssp245',
    ssp585: 'ssp585',
    unknownDefaultOpenApi: '11184809'
} as const;
export type MeshResultSetOptionsScenariosEnum = typeof MeshResultSetOptionsScenariosEnum[keyof typeof MeshResultSetOptionsScenariosEnum];

/**
 * @export
 */
export const MeshResultSetOptionsYearsEnum = {
    NUMBER_1995: 1995,
    NUMBER_2020: 2020,
    NUMBER_2025: 2025,
    NUMBER_2030: 2030,
    NUMBER_2035: 2035,
    NUMBER_2040: 2040,
    NUMBER_2045: 2045,
    NUMBER_2050: 2050,
    NUMBER_2055: 2055,
    NUMBER_2060: 2060,
    NUMBER_2065: 2065,
    NUMBER_2070: 2070,
    NUMBER_2075: 2075,
    NUMBER_2080: 2080,
    NUMBER_2085: 2085,
    NUMBER_2090: 2090,
    NUMBER_2095: 2095,
    NUMBER_2100: 2100,
    NUMBER_unknown_default_open_api: 11184809
} as const;
export type MeshResultSetOptionsYearsEnum = typeof MeshResultSetOptionsYearsEnum[keyof typeof MeshResultSetOptionsYearsEnum];

/**
 * @export
 */
export const MeshResultSetOptionsTypeEnum = {
    mesh: 'mesh',
    unknownDefaultOpenApi: '11184809'
} as const;
export type MeshResultSetOptionsTypeEnum = typeof MeshResultSetOptionsTypeEnum[keyof typeof MeshResultSetOptionsTypeEnum];